+++ modules/apps/export-import/export-import-test/test/integration/com/liferay/export/import/ExportImportHelperUtilTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.exportimport;
+package com.liferay.export.import;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.journal.test.util.JournalTestUtil;
