+++ modules/apps/export-import/export-import-test/test/integration/com/liferay/export/import/dependencies/dl_references.txt
