+++ modules/apps/export-import/export-import-test/test/integration/com/liferay/export/import/dependencies/layout_references.txt
